COM-3273: move function in composable
